+++ HAFoodWeb/Models/CartRequest.cs
@@ -5,9 +5,10 @@
         public long variant_Id { get; set; }
         public int quantity { get; set; }
         public string name_Variant { get; set; }
-        public double? price_Variant { get; set; }
+        public decimal? price_Variant { get; set; } // decimal?
         public string image_Variant { get; set; }
     }
+
 
     public class CartUpdateQtyRequest
     {
